[NEW] Check Underline
This is not a WCAG 2.1 requirement.

This warning is to discourage the use of underlined text as it is easily confused with links.

This condition checks for underlined text created with <u> markup and css styles within the following elements: h1, h2, h3, h4, h5, h6, p, li, span, blockquote.